refact(): rename the Id application (#66)
