+++ app/src/main/java/org/flyve/mdm/agent/data/DataStorage.java
@@ -25,7 +25,7 @@
  * ------------------------------------------------------------------------------
  */
 
-package com.teclib.data;
+package org.flyve.mdm.agent.data;
 
 import android.content.Context;
 import android.content.SharedPreferences;
