+++ app/src/main/java/org/flyve/mdm/agent/security/FlyveAdminReceiver.java
@@ -1,12 +1,13 @@
-package com.teclib.security;
+package org.flyve.mdm.agent.security;
 
-import android.app.admin.DevicePolicyManager;
+import android.app.admin.DeviceAdminReceiver;
 import android.content.Context;
+import android.content.Intent;
 
 /*
  *   Copyright © 2017 Teclib. All rights reserved.
  *
- *   This file is part of flyve-mdm-android
+ *   This file is part of flyve-mdm-android-agent
  *
  * flyve-mdm-android is a subproject of Flyve MDM. Flyve MDM is a mobile
  * device management software.
@@ -21,28 +22,17 @@ import android.content.Context;
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  * ------------------------------------------------------------------------------
- * @author    rafaelhernandez
+ * @author    Rafael Hernandez
  * @date      4/7/17
  * @copyright Copyright © 2017 Teclib. All rights reserved.
  * @license   GPLv3 https://www.gnu.org/licenses/gpl-3.0.html
- * @link      https://github.com/flyve-mdm/flyve-mdm-android
+ * @link      https://github.com/flyve-mdm/flyve-mdm-android-agent
  * @link      https://flyve-mdm.com
  * ------------------------------------------------------------------------------
  */
-public class FlyveDeviceAdminUtils {
-
-    private DevicePolicyManager mDPM;
-
-    public FlyveDeviceAdminUtils(Context context) {
-        mDPM = (DevicePolicyManager) context.getSystemService(Context.DEVICE_POLICY_SERVICE);
+public class FlyveAdminReceiver extends DeviceAdminReceiver {
+    @Override
+    public void onEnabled(Context context, Intent intent) {
+        super.onEnabled(context, intent);
     }
-
-    /**
-     * Erase all data of the device
-     */
-    public void wipe() {
-        mDPM.wipeData(0);
-    }
-
-
 }
