+++ app/src/main/java/org/flyve/mdm/agent/utils/GPSTracker.java
@@ -25,7 +25,7 @@
  * ------------------------------------------------------------------------------
  */
 
-package com.teclib.utils;
+package org.flyve.mdm.agent.utils;
 
 import android.content.Context;
 import android.location.Location;
