+++ app/src/main/java/org/flyve/mdm/agent/utils/Helpers.java
@@ -25,7 +25,7 @@
  * ------------------------------------------------------------------------------
  */
 
-package com.teclib.utils;
+package org.flyve.mdm.agent.utils;
 
 import android.os.Build;
 import android.util.Base64;
